Read attachment data outside the transaction
Previously the data was read from the parser in the transaction. If the
transaction had to retry, for example, because of a conflict, the parser would
have been drained and exited resulting the request failing with a 500
"mp_parser noproc" error.

Since FDB cannot handle transactions larger than 10MB opt to read the
attachment data into memory first, before the transaction starts.