Add fold_jobs/4 and pending_count/2,3 to couch_jobs API
fold_jobs/4 is a generalized folding API which can be used to process all the
jobs of a particular type.

get_pending_count/2,3 can be used to get the count of pending jobs. It takes
the same options as accept, including a limit and `max_sched_time`. Just like
accept it reads the range of pending jobs as a snapshot to avoid generating
read conflicts.